requirements.txt
python-dotenv
Flask
psycopg2-binary
sentence-transformers
PyPDF2
python-docx
transformers 
torch 
flask-cors
redis